+++ src/include/file_helper.h
@@ -27,19 +27,19 @@ public:
     /* Opens a file on specified mode */
     virtual void Open() {}
     /* Reads from file into buffer */
-    virtual int Read(char *buffer, int length) {return -1;}
+    virtual int Read(char *buffer, size_t length) {return -1;}
     /* Writes buffer into file */
-    virtual int Write(char *buffer, int length) {return -1;}
+    virtual int Write(char *buffer, size_t length) {return -1;}
     /* Write Data into file */
-    virtual int WriteData(char *buffer, int length) {return -1;}
+    virtual int WriteData(char *buffer, size_t length) {return -1;}
     /* Write and Sync */
-    virtual int Flush(char *buffer, int length) {return -1;}
+    virtual int Flush(char *buffer, size_t length) {return -1;}
     /* Write Data and Sync it */
-    virtual int FlushData(char *buffer, int length) {return -1;}
+    virtual int FlushData(char *buffer, size_t length) {return -1;}
     /* Append */
     virtual int Append(char *buffer, int legnt) {return -1;}
     /* Seeks to position */
-    virtual void Seek(int offset) {}
+    virtual void Seek(uint64_t offset) {}
     /* */
     virtual uint32_t GetNextLogSize() {return 0;}
     /* Closes the file */
